prevent tick timer from restarting every render
previously, when the state was updated the timer was restarted.
since a tick is not processed until the timer completed a second,
this had the result of never processing any ticks as long as the
state was updating faster than the timer could complete -
for example, if the buzz buzz buzz button was being mashed.

now, the timer interval is never cleaned up
(i dont think it ever needs to be?) so it can just be started once

additionally, we shoud be passing functions into the dispatch functions
that utilize the previous value to prevent any further fuckery