restart timer on update of `bees`
fixes bug introduced in last commit:
previously, because the timer was no longer being restarted, the
amount by which `honey` was incremented did not change with `bees`.
apparently, the `processTick` function does not retain access to any
update to the state, so it would only increment by the value that it
should have at page load (refreshing the page would refresh how much
honey ticked up by).

now, the timer is restarted when the `bees` variable is updated. the
limitation to this approach is that continually changing `bees` quicker
than the timer can lapse will result in no ticks being processed (the
same issue "fixed" by the previous commit).

what is needed to completely alleviate the issue is a way to set the
`processTick` function on an interval that does not restart on re-render
while retaining updated values from the state.